+++ internal/topic/topicwriterinternal/writer_reconnector.go
@@ -569,6 +569,8 @@ func (w *WriterReconnector) WaitInit(ctx context.Context) (info InitialInfo, err
 		return InitialInfo{}, ctx.Err()
 	case <-w.initDoneCh:
 		return w.initInfo, nil
+	case <-w.background.Done():
+		return InitialInfo{}, w.background.CloseReason()
 	}
 }